updates to path planner comments and structure
